AI-ML-Project 01 Exam Helper AI/requirements.txt
streamlit==1.37.0
Pillow==10.4.0
google-generativeai==0.7.2
python-dotenv==1.0.1
gTTS==2.5.3